Split useInnerWatch/watching-multiple-stores.spec
